refactor(profugo): extraer interfaz Profugo y renombrar clase a ProfugoBase
- Se creó la interfaz Profugo para unificar el contrato de comportamiento
- Se renombró la clase original a ProfugoBase para seguir convención Java
- Se actualizaron tests para utilizar ProfugoBase
